swift-tool-version:5.3, add crypto-kit and FoundationNetworking for Linux
