+++ Package.swift
@@ -1,4 +1,4 @@
-// swift-tools-version:5.2
+// swift-tools-version:5.3
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
@@ -36,11 +36,16 @@ let package = Package(
     .package(
       url: "https://github.com/stencilproject/Stencil.git",
       .upToNextMinor(from: "0.14.0")),
+    .package(
+      url: "https://github.com/apple/swift-crypto.git",
+      .upToNextMinor(from: "1.1.2")),
     ],
     targets: [
       .target(
         name: "ApolloCore",
-        dependencies: []),
+        dependencies: [
+          .product(name: "Crypto", package: "swift-crypto", condition: .when(platforms: [.linux]))
+        ]),
     .target(
       name: "Apollo",
       dependencies: [
