+++ Sources/Apollo/NetworkTransport.swift
@@ -1,4 +1,7 @@
 import Foundation
+#if canImport(FoundationNetworking)
+import FoundationNetworking
+#endif
 
 /// A network transport is responsible for sending GraphQL operations to a server.
 public protocol NetworkTransport: class {
