+++ Sources/Apollo/URLSessionClient.swift
@@ -1,4 +1,7 @@
 import Foundation
+#if canImport(FoundationNetworking)
+import FoundationNetworking
+#endif
 #if !COCOAPODS
 import ApolloCore
 #endif
