+++ Sources/ApolloCodegenLib/CLIDownloader.swift
@@ -1,4 +1,7 @@
 import Foundation
+#if canImport(FoundationNetworking)
+import FoundationNetworking
+#endif
 
 // Only available on macOS
 #if os(macOS)
