+++ Sources/ApolloCore/String+SHA.swift
@@ -1,5 +1,4 @@
 import Foundation
-import CommonCrypto
 
 extension String: ApolloCompatible {}
 
@@ -7,16 +6,6 @@ extension ApolloExtension where Base == String {
 
   /// The SHA256 hash of the current string.
   public var sha256Hash: String {
-    let data = base.data(using: .utf8)!
-    var hash = [UInt8](repeating: 0, count: Int(CC_SHA256_DIGEST_LENGTH))
-    data.withUnsafeBytes {
-      _ = CC_SHA256($0.baseAddress, CC_LONG(data.count), &hash)
-    }
-
-    var hashString = ""
-    for byte in hash {
-      hashString += String(format:"%02x", UInt8(byte))
-    }
-    return hashString
+    return Crypt.sha256(base)
   }
 }
