+++ Sources/ApolloTestSupport/MockNetworkTransport.swift
@@ -1,5 +1,9 @@
 @testable import Apollo
 import Dispatch
+import Foundation
+#if canImport(FoundationNetworking)
+import FoundationNetworking
+#endif
 
 public final class MockNetworkTransport: RequestChainNetworkTransport {
