+++ Sources/ApolloTestSupport/MockURLSession.swift
@@ -6,6 +6,9 @@
 //
 
 import Foundation
+#if canImport(FoundationNetworking)
+import FoundationNetworking
+#endif
 import Apollo
 import ApolloCore
